Update recent activities to follow new pattern and remove un-used emptyComponent.jsx file
