+++ frontend/src/features/Activities/ActivityAccordion/ActivityAccordionSummary.jsx
@@ -1,0 +1,14 @@
+import { ExpandMoreRounded } from '@mui/icons-material';
+import { AccordionSummary, Stack, Typography } from '@mui/material';
+import { capitalizeFirstLetter } from '../../common/utils';
+
+export default function ActivityAccordionSummary({ title, label, prefix }) {
+  return (
+    <AccordionSummary expandIcon={<ExpandMoreRounded />}>
+      <Stack>
+        <Typography variant="h6">{title}</Typography>
+        <Typography variant="caption">{`${capitalizeFirstLetter(label)} ${prefix}`}</Typography>
+      </Stack>
+    </AccordionSummary>
+  );
+}
