+++ frontend/src/features/common/router.jsx
@@ -12,7 +12,7 @@ const EditInventory = lazy(() => import('../InventoryList/EditInventory'));
 const CategoryList = lazy(() => import('../Categories/CategoryList'));
 const CategoryItem = lazy(() => import('../Categories/CategoryItem'));
 const MaintenanceItem = lazy(() => import('../Maintenance/MaintenanceItem'));
-const RecentList = lazy(() => import('../Activities/RecentList'));
+const RecentActivityList = lazy(() => import('../Activities/RecentActivityList'));
 
 export const router = createBrowserRouter([
   {
@@ -61,7 +61,7 @@ export const router = createBrowserRouter([
       },
       {
         path: 'recent/activities',
-        element: <RecentList />,
+        element: <RecentActivityList />,
       },
     ],
   },
